+++ src/engin/__init__.py
@@ -3,12 +3,11 @@ from engin._assembler import Assembler
 from engin._block import Block, invoke, provide
 from engin._dependency import Entrypoint, Invoke, Provide, Supply
 from engin._engin import Engin, Option
-from engin._exceptions import AssemblyError
+from engin._exceptions import ProviderError
 from engin._lifecycle import Lifecycle
 
 __all__ = [
     "Assembler",
-    "AssemblyError",
     "Block",
     "Engin",
     "Entrypoint",
@@ -16,6 +15,7 @@ __all__ = [
     "Lifecycle",
     "Option",
     "Provide",
+    "ProviderError",
     "Supply",
     "ext",
     "invoke",