+++ CHANGELOG.md
@@ -6,6 +6,18 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## [0.0.5] - 2025-01-29
+
+### Added
+
+- Docstrings for every public class, method and function.
+
+### Changed
+
+- AssemblyError has been renamed to ProviderError.
+- Lifecycle now supports synchronous Context Managers.
+
+
 ## [0.0.4] - 2025-01-27
 
 ### Changed